updated the images and css

+++ admin/delete_message.php
@@ -1,11 +1,17 @@
-<?php 
+<?php
 include('../connect.php');
-$id = $_POST['id'];
-if(isset($_POST['submitButton'])){
-		$deleteQuery = "DELETE FROM message WHERE id=$id";
-		mysqli_query($connection, $deleteQuery);
-        header("Location: message.php");
-        // echo "<meta http-equiv='refresh' content='0'>";
-	}
 
+if (isset($_POST['submitButton']) && isset($_POST['id'])) {
+    $id = intval($_POST['id']); // Sanitize input
+
+    $deleteQuery = "DELETE FROM message WHERE id = $id";
+    if (mysqli_query($connection, $deleteQuery)) {
+        header("Location: message.php"); // Redirect after successful deletion
+        exit();
+    } else {
+        echo "Error deleting message: " . mysqli_error($connection);
+    }
+} else {
+    echo "Invalid request.";
+}
 ?>
